Add function result to identifiers and dependencies
Added the function result to the quoted and unquoted identifiers, as well as to the dependencies. This additional parameter was included in multiple functions and SQL queries, and it allows to better identify and manage dependencies in the 'schemainspect' Python module.

(cherry picked from commit 91fe044711901cac6bf3c6e41bde116f72417412)